+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/asbuilt/mapper/AssetAsBuiltFieldMapper.java
@@ -39,8 +39,6 @@ public class AssetAsBuiltFieldMapper extends BaseRequestFieldMapper {
             Map.entry("idShort", "idShort"),
             Map.entry("owner", "owner"),
             Map.entry("classification", "classification"),
-            Map.entry("underInvestigation", "inInvestigation"),
-            Map.entry("activeAlert", "activeAlert"),
             Map.entry("qualityType", "qualityType"),
             Map.entry("manufacturerPartId", "manufacturerPartId"),
             Map.entry("manufacturerName", "manufacturerName"),