+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/importpoc/PublishService.java
@@ -16,12 +16,11 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
+package org.eclipse.tractusx.traceability.assets.application.importpoc;
 
-package org.eclipse.tractusx.traceability.submodel.infrastructure.reposotory;
+import java.util.List;
 
-import org.eclipse.tractusx.traceability.submodel.infrastructure.model.SubmodelPayloadEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
+public interface PublishService {
 
-public interface JpaSubmodelPayloadRepository extends JpaRepository<SubmodelPayloadEntity, String> {
-
+    void publishAssets(String policyId, List<String> assetIds);
 }