+++ CHANGELOG.md
@@ -8,6 +8,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [UNRELEASED - DD.MM.YYYY]
 
 ### Added
+
+### Changed
+
+### Removed
+
+## [10.2.0 - 22.01.2024]
+
+### Added
 - Import Data Service for data provisioning
 - Added UI to publish own assets
 - Sequence Diagrams for Data Provisioning Flow
@@ -16,6 +24,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added bpn validator to bpn edc mapping
 - Added new /irs/job/callback endpoint to handle irs job finished callback
 - Added Asset import info to parts table and parts detail view
+- SubmodelServerService for calling submodel server with feign client
+- Added GET /policies endpoint to retrieve accepted policies
+- Added POST assets/publish endpoint to publish transient assets
 
 ### Changed
 - Fixed security findings
@@ -31,6 +42,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Upgraded irs version from 6.12.0 to 6.13.0
 - Switched from OAuth2.0 to API Key authentication for IRS API Requests
 - switched from json-schema-friend:0.12.3 to json-schema-validator:5.4.0 for import file validation
+- Moved logic for active notifications to frontend
 - Updated com.github.spotbugs:spotbugs-maven-plugin from 4.7.3.0 to 4.8.3.0
 - Updated actions/github-script from 5 to 7
 - Updated org.apache.maven.plugins:maven-surefire-plugin from 3.1.2 to 3.2.5